check to see if user can run/delete #630
This makes it so only people with a certain page level access can run
and delete runs. The following section will need to be inserted in
config.php for this to work.

```
# Require username/password, can set min level to 0 so nobody can
run/delete.
# 4 = viewer
# 3 = creator
# 2 = manager
# 1 = administrator
$authentication=true;
$min_run_level=2;
$min_delete_level=2;
```
