+++ web/01-introduction.php
@@ -12,13 +12,18 @@
 require("common.php");
 
 if ($authentication) {
-	open_database();
-	if (!check_login()) {
-		header( "Location: index.php");
-		close_database();
-		exit;
-	}
-	close_database();
+    open_database();
+    if (!check_login()) {
+        header( "Location: index.php");
+        close_database();
+        exit;
+    }
+    if (get_page_acccess_level() > $min_run_level) {
+        header( "Location: history.php");
+        close_database();
+        exit;
+    }
+    close_database();
 }
 
 ?>
@@ -35,39 +40,39 @@ if ($authentication) {
         $("#error").html("");
     }
         
-	function prevStep() {
-		$("#formprev").submit();
-		}
+    function prevStep() {
+        $("#formprev").submit();
+        }
 
-	function nextStep() {
-		console.log($("#formnext"));
-		$("#formnext").submit();
-	}
+    function nextStep() {
+        console.log($("#formnext"));
+        $("#formnext").submit();
+    }
 </script>
 </head>
 <body>
 <div id="wrap">
-	<div id="stylized">
-		<form id="formprev" method="POST" action="history.php">
-		</form>
-		<form id="formnext" method="POST" action="02-modelsite.php">
-			<h1>Introduction</h1>
-			<p>Below you will find the buttons to step through the
-			workflow creation process.</p>
+    <div id="stylized">
+        <form id="formprev" method="POST" action="history.php">
+        </form>
+        <form id="formnext" method="POST" action="02-modelsite.php">
+            <h1>Introduction</h1>
+            <p>Below you will find the buttons to step through the
+            workflow creation process.</p>
 
-			<label>Offline mode:</label>
-			<span id="error" class="small">This will disable Google Maps</span>
-			<input name="offline" id="offline" style="align: left" type="checkbox" value="offline">
-			<div class="spacer"></div>
+            <label>Offline mode:</label>
+            <span id="error" class="small">This will disable Google Maps</span>
+            <input name="offline" id="offline" style="align: left" type="checkbox" value="offline">
+            <div class="spacer"></div>
 
-			<p></p>
-			<label>Workflow</label>
-			<span id="error" class="small">&nbsp;</span>
-			<input id="prev" type="button" value="History" onclick="prevStep();" />
-			<input id="next" type="button" value="Next" onclick="nextStep();" />
-			
-			<div class="spacer"></div>
-		</form>
+            <p></p>
+            <label>Workflow</label>
+            <span id="error" class="small">&nbsp;</span>
+            <input id="prev" type="button" value="History" onclick="prevStep();" />
+            <input id="next" type="button" value="Next" onclick="nextStep();" />
+            
+            <div class="spacer"></div>
+        </form>
 <?php
   if (check_login()) {
     echo "<p></p>";
@@ -75,32 +80,32 @@ if ($authentication) {
     echo "<a href=\"index.php?logout\" id=\"logout\">logout</a>";
   }
 ?>    
-	</div>
-	<div id="output">
-		<h1>Introduction</h1>
-		<p>The following pages will guide you through setting up a
-		PEcAn worklflow. You will be able to always go back to a
-		previous step to change inputs. However once the model is
-		running it will continue to run until it finishes. You will
-		be able to use the history button to jump to existing 
-		executions of PEcAn.</p>
-		<p>The following webpages will help to setup the PEcAn
-		workflow. You will be asked the following questions:</p>
-		<ol>
-		<li><b>Host and Model</b> You will first select the host to
-		run the workflow on as well as the model to be exectuted.</li>
-		<li><b>Site</b> The next step is to select the site where
-		the model should be run for.</li>
-		<li><b>Model Parameters</b> Based on the site some final
-		parameters for the model will need to be selected.</li>
-		<li><b>Model Execution</b> Once all variables are selected
-		PEcAn will execute the workflow.</li>
-		<li><b>Results</b> After execution of the PEcAn workflow you
-		will be presented with a page showing the results of the
-		PEcAn workflow.</li> 
-		</ol>
-	</div>
-	<div id="footer"><?php echo get_footer(); ?></div>
+    </div>
+    <div id="output">
+        <h1>Introduction</h1>
+        <p>The following pages will guide you through setting up a
+        PEcAn worklflow. You will be able to always go back to a
+        previous step to change inputs. However once the model is
+        running it will continue to run until it finishes. You will
+        be able to use the history button to jump to existing 
+        executions of PEcAn.</p>
+        <p>The following webpages will help to setup the PEcAn
+        workflow. You will be asked the following questions:</p>
+        <ol>
+        <li><b>Host and Model</b> You will first select the host to
+        run the workflow on as well as the model to be exectuted.</li>
+        <li><b>Site</b> The next step is to select the site where
+        the model should be run for.</li>
+        <li><b>Model Parameters</b> Based on the site some final
+        parameters for the model will need to be selected.</li>
+        <li><b>Model Execution</b> Once all variables are selected
+        PEcAn will execute the workflow.</li>
+        <li><b>Results</b> After execution of the PEcAn workflow you
+        will be presented with a page showing the results of the
+        PEcAn workflow.</li> 
+        </ol>
+    </div>
+    <div id="footer"><?php echo get_footer(); ?></div>
 </div>
 </body>
 </html>
