+++ web/config.example.php
@@ -23,8 +23,14 @@ $browndog_password="";
 # R binary
 $Rbinary="/usr/bin/R";
 
-# Require username/password
+# Require username/password, can set min level to 0 so nobody can run/delete.
+# 4 = viewer
+# 3 = creator
+# 2 = manager
+# 1 = administrator
 $authentication=false;
+$min_run_level=2;
+$min_delete_level=2;
 
 # Used for authentication, needs to be same as ruby
 $REST_AUTH_SITE_KEY="thisisnotasecret";
