+++ web/dataset.php
@@ -17,6 +17,11 @@ if ($authentication) {
 		header('HTTP/1.1 403 Unauthorized');
 		exit;
 	}
+    if (get_page_acccess_level() > $min_run_level) {
+        header( "Location: history.php");
+        close_database();
+        exit;
+    }
 }
 
 // runid
