+++ web/sites.php
@@ -16,10 +16,15 @@ require("common.php");
 open_database();
 if ($authentication) {
 	if (!check_login()) {
-		close_database();
 		header('HTTP/1.1 403 Unauthorized');
+    close_database();
 		exit;
 	}
+  if (get_page_acccess_level() > $min_run_level) {
+    header('HTTP/1.1 403 Unauthorized');
+    close_database();
+    exit;
+  }
 }
 
 // Start XML file, create parent node
